finished improvements

+++ main.py
@@ -7,8 +7,10 @@ import os
 database.create_connection("data")
 database.create_table()
 
-with open('.env', 'w') as file:
-    file.write('LEAGUE_PATH=C:\Riot Games\League of Legends\LeagueClient.exe')
+if not os.path.exists(".env"):
+    with open('.env', 'w') as file:
+        file.write('LEAGUE_PATH=C:\Riot Games\League of Legends\LeagueClient.exe')
+
 
 if not os.path.exists("secret.key"):
     acc.Encryption.generate_key()
